requirements/django.txt
-r base.txt

# Django and related packages
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
django-environ==0.11.2 